+++ parquet/src/encodings/encoding/dict_encoder.rs
@@ -123,12 +123,12 @@ impl<T: DataType> DictEncoder<T> {
     pub fn write_indices(&mut self) -> Result<ByteBufferPtr> {
         let buffer_len = self.estimated_data_encoded_size();
         let mut buffer = Vec::with_capacity(buffer_len);
-        buffer.push(self.bit_width() as u8);
+        buffer.push(self.bit_width());
 
         // Write bit width in the first byte
         let mut encoder = RleEncoder::new_from_buf(self.bit_width(), buffer);
         for index in &self.indices {
-            encoder.put(*index as u64)
+            encoder.put(*index)
         }
         self.indices.clear();
         Ok(ByteBufferPtr::new(encoder.consume()))
